requirements.txt
streamlit
PyPDF2
openai
chromadb
langchain
langchain-groq
python-dotenv
tiktoken
sentence-transformers
langchain_community
langchain-huggingface